+++ src/extensions/AssetView.ts
@@ -81,7 +81,7 @@ export class AssetView
     rootItem: AssetViewItem;
     path: string[] = [];
     rootPath: string = '';
-    selectedItem: AssetViewItem | undefined = undefined;
+    selectedItems: AssetViewItem[] = [];
     allowedTypes: any;
     searchValue: string = '';
     filter: string = 'None';
@@ -94,7 +94,8 @@ export class AssetView
     skipPreview: boolean = false;
     useNativeTitle: boolean = false;
     onlyFolders: boolean = true;
-    allowMultipleSelection: boolean = false;
+    allowMultipleSelection: boolean = true;
+    allowItemCheck: boolean = false;
     previewActions: any[] = [];
     contextMenu: any[] = [];
     itemContextMenuOptions: any = null;
@@ -148,6 +149,7 @@ export class AssetView
         this.useNativeTitle = options.useNativeTitle ?? this.useNativeTitle;
         this.onlyFolders = options.onlyFolders ?? this.onlyFolders;
         this.allowMultipleSelection = options.allowMultipleSelection ?? this.allowMultipleSelection;
+        this.allowItemCheck = options.allowItemCheck ?? this.allowItemCheck;
         this.previewActions = options.previewActions ?? [];
         this.itemContextMenuOptions = options.itemContextMenuOptions;
         this.gridScale = options.gridScale ?? this.gridScale;
@@ -283,9 +285,9 @@ export class AssetView
         {
             let desc = document.createElement( 'span' );
             desc.className = 'lexitemdesc';
-            desc.id = `floatingTitle_${metadata.uid}`;
+            desc.id = LX.getSupportedDOMName( `floatingTitle_${metadata.uid}` );
             desc.innerHTML = `File: ${item.id}<br>Type: ${type}`;
-            LX.insertChildAtIndex( this.content, desc, childIndex ? childIndex + 1 : undefined );
+            LX.insertChildAtIndex( this.content, desc, childIndex !== undefined ? childIndex + 1 : undefined );
 
             itemEl.addEventListener( 'mousemove', ( e: MouseEvent ) => {
                 if ( !isGridLayout )
@@ -322,7 +324,7 @@ export class AssetView
             itemEl.title = type + ': ' + item.id;
         }
 
-        if ( this.allowMultipleSelection )
+        if ( this.allowItemCheck )
         {
             let checkbox = document.createElement( 'input' );
             checkbox.type = 'checkbox';
@@ -440,13 +442,14 @@ export class AssetView
 
             if ( !isDoubleClick )
             {
-                if ( !e.shiftKey )
+                if ( !e.shiftKey || !that.allowMultipleSelection )
                 {
                     that.content.querySelectorAll( '.lexassetitem' ).forEach( ( i ) => i.classList.remove( 'selected' ) );
+                    that.selectedItems.length = 0;
                 }
 
                 this.classList.add( 'selected' );
-                that.selectedItem = item;
+                that.selectedItems.push( item );
 
                 if ( !that.skipPreview )
                 {
@@ -489,11 +492,11 @@ export class AssetView
             e.stopImmediatePropagation();
             e.stopPropagation();
 
-            const multiple = that.content.querySelectorAll( '.selected' ).length;
+            const multipleSelection = that.selectedItems.length > 1;
 
             const options: any[] = [
                 {
-                    name: ( multiple > 1 ) ? ( multiple + ' selected' ) : item.id,
+                    name: multipleSelection ? ( `${that.selectedItems.length} selected` ) : item.id,
                     icon: LX.makeIcon( 'CircleSmall', { svgClass: `fill-current text-${typeColor}` } ),
                     className: 'text-sm',
                     disabled: true
@@ -501,35 +504,44 @@ export class AssetView
                 null
             ];
 
-            if ( multiple <= 1 )
+            // By now, allow with none/single selected items
+            if ( !multipleSelection )
             {
                 options.push( { name: 'Rename', icon: 'TextCursor', callback: that._renameItemPopover.bind( that, item ) } );
             }
 
-            if ( !isFolder )
+            // By now, allow with none/single selected items
+            if ( !isFolder && !multipleSelection )
             {
                 options.push( { name: 'Clone', icon: 'Copy', callback: that._requestCloneItem.bind( that, item ) } );
             }
 
-            options.push( { name: 'Move', icon: 'FolderInput', callback: () => that._moveItem( item ) } );
+            // By now, allow with none/single selected items
+            if ( !multipleSelection )
+            {
+                options.push( { name: 'Move', icon: 'FolderInput', callback: () => that._moveItem( item ) } );
+            }
 
-            if ( type == 'Script' && LX.has( 'CodeEditor' ) )
+            // By now, allow with none/single selected items
+            if ( !multipleSelection && type == 'Script' && LX.has( 'CodeEditor' ) )
             {
                 options.push( { name: 'Open in Editor', icon: 'Code', callback: that._openScriptInEditor.bind( that, item ) } );
             }
 
             if ( that.itemContextMenuOptions )
             {
-                options.push( null );
+                if ( options.length > 2 ) options.push( null );
 
                 for ( let o of that.itemContextMenuOptions )
                 {
                     if ( !o.name || !o.callback ) continue;
-                    options.push( { name: o.name, icon: o.icon, callback: o.callback?.bind( that, item ) } );
+                    options.push( { name: o.name, icon: o.icon,
+                        callback: o.callback?.bind( that, multipleSelection ? that.selectedItems : [ item ] ) } );
                 }
             }
 
-            options.push( null, { name: 'Delete', icon: 'Trash2', className: 'destructive', callback: that._requestDeleteItem.bind( that, item ) } );
+            options.push( null, { name: 'Delete', icon: 'Trash2', className: 'destructive',
+                callback: that._requestDeleteItem.bind( that, multipleSelection ? that.selectedItems : [ item ] ) } );
 
             LX.addClass( that.contentPanel.root, 'pointer-events-none' );
 
@@ -575,7 +587,8 @@ export class AssetView
                 e.dataTransfer.effectAllowed = 'move';
             }
 
-            const desc: HTMLElement | null = that.content.querySelector( `#floatingTitle_${metadata.uid}` );
+            const domName: string = LX.getSupportedDOMName( `floatingTitle_${metadata.uid}` );
+            const desc: HTMLElement | null = that.content.querySelector( `#${domName}` );
             if ( desc ) desc.style.display = 'none';
         }, false );
 
@@ -616,7 +629,8 @@ export class AssetView
         itemEl.addEventListener( 'mouseenter', ( e: MouseEvent ) => {
             if ( !that.useNativeTitle && isGridLayout )
             {
-                const desc: HTMLElement | null = that.content.querySelector( `#floatingTitle_${metadata.uid}` );
+                const domName: string = LX.getSupportedDOMName( `floatingTitle_${metadata.uid}` );
+                const desc: HTMLElement | null = that.content.querySelector( `#${domName}` );
                 if ( desc ) desc.style.display = 'unset';
             }
 
@@ -631,7 +645,8 @@ export class AssetView
             if ( !that.useNativeTitle && isGridLayout )
             {
                 setTimeout( () => {
-                    const desc: HTMLElement | null = that.content.querySelector( `#floatingTitle_${metadata.uid}` );
+                    const domName: string = LX.getSupportedDOMName( `floatingTitle_${metadata.uid}` );
+                    const desc: HTMLElement | null = that.content.querySelector( `#${domName}` );
                     if ( desc ) desc.style.display = 'none';
                 }, 100 );
             }
@@ -808,7 +823,7 @@ export class AssetView
                     dom?.classList.add( 'selected' );
                 }
 
-                this.selectedItem = node;
+                this.selectedItems = [ node ];
             }
         } );
 
@@ -1004,6 +1019,7 @@ export class AssetView
         this.content.addEventListener( 'click', function()
         {
             this.querySelectorAll( '.lexassetitem' ).forEach( ( i ) => i.classList.remove( 'selected' ) );
+            that.selectedItems.length = 0;
         } );
 
         this.content.addEventListener( 'contextmenu', function( e )
@@ -1297,7 +1313,7 @@ export class AssetView
         }
 
         const child = this.currentData[0];
-        const sameFolder = child?.parent?.id === folderItem.id;
+        const sameFolder = child?.parent?.metadata?.uid === folderItem.metadata?.uid;
 
         if ( storeCurrent )
         {
@@ -1332,7 +1348,20 @@ export class AssetView
         {
             this._processData( this.data );
             this._refreshContent();
-            this.tree?.select( this.currentFolder.id );
+
+            // Get path to avoid same id issues
+
+            let path = `${this.currentFolder.id}/`;
+            let parent = this.currentFolder.parent;
+
+            while ( parent && parent.id !== '/' )
+            {
+                path += `${parent.id}/`;
+                parent = parent.parent;
+            }
+
+            const parentsPath = path.split( '/' ).filter( Boolean ).reverse();
+            this.tree?.select( undefined, parentsPath );
         }
 
         this._updatePath();
@@ -1368,16 +1397,16 @@ export class AssetView
         return true;
     }
 
-    _requestDeleteItem( item: AssetViewItem )
+    _requestDeleteItem( items: AssetViewItem[] )
     {
         const onBeforeDelete = this._callbacks['beforeDelete'];
         const onDelete = this._callbacks['delete'];
 
         const resolve = ( ...args: any[] ) => {
-            this._deleteItem( item );
+            items.forEach( ( item ) => this._deleteItem( item ) );
             const event: AssetViewEvent = {
                 type: 'delete',
-                items: [ item ],
+                items,
                 userInitiated: true
             };
             if ( onDelete ) onDelete( event, ...args );
@@ -1387,7 +1416,7 @@ export class AssetView
         {
             const event: AssetViewEvent = {
                 type: 'delete',
-                items: [ item ],
+                items,
                 userInitiated: true
             };
             onBeforeDelete( event, resolve );
@@ -1540,7 +1569,7 @@ export class AssetView
 
             bcContainer.innerHTML = '';
             bcContainer.appendChild( LX.makeBreadcrumb( path.reverse().map( ( p ) => {
-                return { title: p };
+                return { name: p };
             } ), {
                 maxItems: 4,
                 separatorIcon: 'ChevronRight'
@@ -1760,12 +1789,13 @@ export class AssetView
         if ( item.domEl )
         {
             const wasSelected = LX.hasClass( item.domEl, 'selected' );
-            const hoverTitle: HTMLElement | null = this.content.querySelector(
-                `#floatingTitle_${item.id.replace( /\s/g, '_' ).replaceAll( '.', '_' )}`
-            );
+            const hoverTitleDomName: string = LX.getSupportedDOMName( `floatingTitle_${item.metadata.uid}` );
+            const hoverTitle: HTMLElement | null = this.content.querySelector( `#${hoverTitleDomName}` );
             if ( hoverTitle ) hoverTitle.remove();
 
             item.domEl?.remove();
+            // Update new name
+            item.id = newName;
             item.domEl = this.addItem( item, idx * 2 );
 
             if ( wasSelected )
@@ -1773,8 +1803,10 @@ export class AssetView
                 this._previewAsset( item );
             }
         }
-
-        item.id = newName;
+        else
+        {
+            item.id = newName;
+        }
 
         this.tree?.refresh();